car_price_predictor/requirements.txt
# Django
Django==5.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0

# Web Scraping
requests==2.31.0
beautifulsoup4==4.12.2

# Data Processing
pandas==2.1.4
numpy==1.26.2

# Machine Learning
scikit-learn==1.3.2

# Task Scheduling
celery==5.3.4
redis==5.0.1
